HR Diversity & retention Report
Human Resources at our telecom client is highly into diversity and inclusion. They’ve been working hard to improve gender balance at the executive management level, but they’re not seeing any progress. They’re reaching out to us for help.

At PwC Switzerland we are often approached by clients seeking support with diversity and inclusion. Companies need a workforce of diverse talents and backgrounds to succeed in an increasingly complex and heterogeneous world. To us, diversity and inclusion are business imperatives, not just nice-to-haves. We aim for all of our teams to feel welcome and appreciated. But actually achieving this and unlocking its potential involves a whole set of practical challenges.

Why is this so?

Think about the importance of strategy, awareness and education, analytics and inspiration. 

Here is a hint: Calculating the following measures could help to define proper KPIs:

# of men
# of women
# of leavers
% employees promoted (FY21)
% of women promoted
% of hires men
% of hires women
% turnover 
Average performance rating: men
Average Performance rating: women